pyproject.toml

[project]
name = "soarm_sim"
version = "0.0.1"
description = "SO Arm simulation with mujoco."
readme = "README.md"

requires-python = ">=3.10"
dependencies = [
  "mujoco",
  "dm-control",
  "numpy",
  "pyzmq>=26",
  "dynamixel-sdk>=3.7.31",
  "pyserial>=3.5",
]
